Moved auth-related tests out of ConnectionTests into their own test suite
No actual changes, just move code around
